make it clear where to use each sanitizer
Shuffle classes and packages around to make the distinction between
validation rules for POJOs (EntityValidator) and admission control rules
on API servers.
