+++ titus-common/src/main/java/com/netflix/titus/common/model/sanitizer/ValidationError.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Netflix, Inc.
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.common.model.validator;
+package com.netflix.titus.common.model.sanitizer;
 
 /**
- * Validation errors are generated by {@link EntityValidator}s.  They indicate various reasons why a Job may be malformed.
+ * Validation errors indicate various reasons why a Job may be malformed.
  */
 public class ValidationError {
     private final String field;
@@ -31,11 +31,7 @@ public class ValidationError {
      */
     public enum Type {
         HARD,
-        SOFT;
-
-        public static Type from(EntityValidatorConfiguration configuration) {
-            return valueOf(configuration.getErrorType().trim().toUpperCase());
-        }
+        SOFT
     }
 
     public ValidationError(String field, String description) {
