+++ titus-common/src/main/java/com/netflix/titus/common/model/sanitizer/internal/DefaultEntitySanitizer.java
@@ -27,18 +27,18 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
 import javax.validation.Validator;
 
 import com.netflix.titus.common.model.sanitizer.EntitySanitizer;
+import com.netflix.titus.common.model.sanitizer.ValidationError;
 import com.netflix.titus.common.model.sanitizer.VerifierMode;
-import com.netflix.titus.common.model.validator.ValidationError;
 import org.springframework.expression.EvaluationContext;
 import org.springframework.expression.spel.support.ReflectiveMethodResolver;
 import org.springframework.expression.spel.support.StandardEvaluationContext;
 
 /**
+ *
  */
 public class DefaultEntitySanitizer implements EntitySanitizer {
