+++ titus-server-gateway/src/main/java/com/netflix/titus/gateway/startup/TitusGatewayModule.java
@@ -52,8 +52,8 @@ import com.netflix.titus.runtime.connector.relocation.noop.NoOpRelocationConnect
 import com.netflix.titus.runtime.connector.titusmaster.TitusMasterConnectorModule;
 import com.netflix.titus.runtime.endpoint.common.EmptyLogStorageInfo;
 import com.netflix.titus.runtime.endpoint.common.LogStorageInfo;
-import com.netflix.titus.runtime.endpoint.validator.JobSecurityValidatorModule;
-import com.netflix.titus.runtime.endpoint.validator.TitusValidatorModule;
+import com.netflix.titus.runtime.endpoint.admission.JobSecurityValidatorModule;
+import com.netflix.titus.runtime.endpoint.admission.TitusValidatorModule;
 
 // Common module dependencies
 // Server dependencies
