+++ titus-server-gateway/src/test/java/com/netflix/titus/gateway/service/v3/internal/JobSecurityValidatorTest.java
@@ -26,9 +26,9 @@ import com.netflix.titus.api.iam.service.IamConnectorException;
 import com.netflix.titus.api.jobmanager.JobAttributes;
 import com.netflix.titus.api.jobmanager.model.job.JobDescriptor;
 import com.netflix.titus.api.jobmanager.model.job.SecurityProfile;
-import com.netflix.titus.common.model.validator.ValidationError;
-import com.netflix.titus.runtime.endpoint.validator.JobIamValidator;
-import com.netflix.titus.runtime.endpoint.validator.JobSecurityValidatorConfiguration;
+import com.netflix.titus.common.model.sanitizer.ValidationError;
+import com.netflix.titus.runtime.endpoint.admission.JobIamValidator;
+import com.netflix.titus.runtime.endpoint.admission.JobSecurityValidatorConfiguration;
 import com.netflix.titus.testkit.model.job.JobDescriptorGenerator;
 import org.junit.Before;
 import org.junit.Test;
