+++ titus-server-runtime/src/main/java/com/netflix/titus/runtime/endpoint/admission/FailJobValidator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Netflix, Inc.
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,37 +14,46 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.api.jobmanager.model.job.validator;
+package com.netflix.titus.runtime.endpoint.admission;
 
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.Set;
+import java.util.UUID;
 
 import com.netflix.titus.api.jobmanager.model.job.JobDescriptor;
-import com.netflix.titus.common.model.validator.EntityValidator;
-import com.netflix.titus.common.model.validator.ValidationError;
+import com.netflix.titus.api.service.TitusServiceException;
+import com.netflix.titus.common.model.sanitizer.ValidationError;
 import reactor.core.publisher.Mono;
 
 /**
- * This validator never completes.  It is used to test validation in the face of unresponsive service calls.
+ * This {@link EntityValidator} implementation always causes validation to fail.  It is used for testing purposes.
  */
-public class NeverJobValidator implements EntityValidator<JobDescriptor> {
+public class FailJobValidator implements EntityValidator<JobDescriptor> {
+    public static final String ERR_FIELD = "fail-field";
+    public static final String ERR_DESCRIPTION = "The FailJobValidator should always fail with a unique error:";
+
     private final ValidationError.Type errorType;
 
-    public NeverJobValidator() {
+    public FailJobValidator() {
         this(ValidationError.Type.HARD);
     }
 
-    public NeverJobValidator(ValidationError.Type errorType) {
+    public FailJobValidator(ValidationError.Type errorType) {
         this.errorType = errorType;
     }
 
     @Override
     public Mono<Set<ValidationError>> validate(JobDescriptor entity) {
-        return Mono.never();
+        final String errorMsg = String.format("%s %s", ERR_DESCRIPTION, UUID.randomUUID().toString());
+        final ValidationError error = new ValidationError(ERR_FIELD, errorMsg);
+
+        return Mono.just(new HashSet<>(Collections.singletonList(error)));
     }
 
     @Override
     public Mono<JobDescriptor> sanitize(JobDescriptor entity) {
-        return Mono.never();
+        return Mono.error(TitusServiceException.invalidArgument(ERR_DESCRIPTION));
     }
 
     @Override
