+++ titus-server-runtime/src/main/java/com/netflix/titus/runtime/endpoint/admission/JobIamValidator.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.netflix.titus.runtime.endpoint.validator;
+package com.netflix.titus.runtime.endpoint.admission;
 
 import java.time.Duration;
 import java.util.Collections;
@@ -30,8 +30,7 @@ import com.netflix.titus.api.iam.service.IamConnectorException;
 import com.netflix.titus.api.jobmanager.JobAttributes;
 import com.netflix.titus.api.jobmanager.model.job.JobDescriptor;
 import com.netflix.titus.api.jobmanager.model.job.JobFunctions;
-import com.netflix.titus.common.model.validator.EntityValidator;
-import com.netflix.titus.common.model.validator.ValidationError;
+import com.netflix.titus.common.model.sanitizer.ValidationError;
 import reactor.core.publisher.Mono;
 
 /**
@@ -89,7 +88,7 @@ public class JobIamValidator implements EntityValidator<JobDescriptor> {
                             new ValidationError(
                                     JobIamValidator.class.getSimpleName(),
                                     throwable.getMessage(),
-                                    getErrorType())));
+                                    configuration.toValidatorErrorType())));
                 });
     }
 
@@ -126,7 +125,7 @@ public class JobIamValidator implements EntityValidator<JobDescriptor> {
 
     @Override
     public ValidationError.Type getErrorType() {
-        return ValidationError.Type.from(configuration);
+        return configuration.toValidatorErrorType();
     }
 
     private boolean isIamArn(String iamRoleName) {
