+++ titus-server-runtime/src/main/java/com/netflix/titus/runtime/endpoint/admission/JobImageValidatorConfiguration.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.runtime.endpoint.validator;
+package com.netflix.titus.runtime.endpoint.admission;
 
 import com.netflix.archaius.api.annotations.Configuration;
 import com.netflix.archaius.api.annotations.DefaultValue;
-import com.netflix.titus.common.model.validator.EntityValidatorConfiguration;
 
 @Configuration(prefix = "titus.validate.job.image")
 public interface JobImageValidatorConfiguration extends EntityValidatorConfiguration {
