+++ titus-server-runtime/src/test/java/com/netflix/titus/runtime/endpoint/admission/TestingCapacityGroupSanitizer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Netflix, Inc.
+ * Copyright 2019 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,23 +14,22 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.api.jobmanager.model.job.validator;
+package com.netflix.titus.runtime.endpoint.admission;
 
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 
 import com.netflix.titus.api.jobmanager.model.job.JobDescriptor;
-import com.netflix.titus.common.model.validator.EntityValidator;
-import com.netflix.titus.common.model.validator.ValidationError;
+import com.netflix.titus.common.model.sanitizer.ValidationError;
 import reactor.core.publisher.Mono;
 
 /**
  * This {@link EntityValidator} implementation ensures a job's capacity group matches a specific
  * string. It is only used for testing purposes.
  */
-public class TestingCapacityGroupSanitizer implements EntityValidator<JobDescriptor> {
-    public final static String desiredCapacityGroup = "desiredCapacityGroup";
+class TestingCapacityGroupSanitizer implements EntityValidator<JobDescriptor> {
+    final static String desiredCapacityGroup = "desiredCapacityGroup";
     private static final String ERR_FIELD = "fail-field";
     private static final String ERR_DESCRIPTION =
             String.format("The job does not have desired capacity group %s", desiredCapacityGroup);
@@ -51,6 +50,7 @@ public class TestingCapacityGroupSanitizer implements EntityValidator<JobDescrip
                 .build());
     }
 
+    @Override
     public ValidationError.Type getErrorType() {
         return ValidationError.Type.HARD;
     }
